+++ packages/graphql/src/schema/query.ts
@@ -38,6 +38,7 @@ import {
   CourseStudentTimeline,
   CourseSummary,
   LeaderboardEntry,
+  StudentAssessmentResults,
   StudentCourse,
 } from './course.js'
 import {
@@ -287,6 +288,9 @@ export const Query = builder.queryType({
           showDependencies: t.arg.boolean({ required: false }),
           multiplier: t.arg.int({ required: false }),
           reviewStatus: t.arg({ type: ReviewStatus, required: false }),
+          isGamificationEnabled: t.arg.boolean({ required: false }),
+          isAssessmentEnabled: t.arg.boolean({ required: false }),
+          isPinProtected: t.arg.boolean({ required: false }),
           sortByType: t.arg({ type: SortByType, required: true }),
           sortByAsc: t.arg.boolean({ required: true }),
           numEntries: t.arg.int({ required: false }),
@@ -889,6 +893,15 @@ export const Query = builder.queryType({
         },
         resolve: async (_, args, ctx) => {
           return await LiveQuizService.getLiveQuizLeaderboard(args, ctx)
+        },
+      }),
+
+      studentAssessmentResults: t.withAuth(asParticipant).field({
+        nullable: true,
+        type: StudentAssessmentResults,
+        args: { courseId: t.arg.string({ required: true }) },
+        resolve: async (_, args, ctx) => {
+          return await CourseService.getStudentAssessmentResults(args, ctx)
         },
       }),